Use block action to simplify templates
This change uses a base template containing block actions such as title
and content. The blocks are filled by combining the base template with
the home and system templates which define title and content. This
simplifies the templates as home and system contain only the necessary
information they need while they "inherit" the rest from the base
template. Additionally, the base template also contains blocks for css
and js. The system template overrides the js block by defining js.

The pageInfo struct is changed to contain an Active field which is
passed to each template to decide which navigation link to mark as
active.