Use a server struct for all the dependencies
This commit adds a server struct which contains all the program's
dependencies such as the temperature and frequency files as well as the
directory of the templates. This allows to change the paths using flags
when the program starts.

The server struct implements ServeHTTP and thus can be used like any
other http.Handler.